Use byte for color in def

+++ Assets/Scripts/Defs/ThingDef.cs
@@ -12,7 +12,7 @@ namespace Assets.Scripts.Defs {
 		[DefaultValue(null)] public string TexPath;
 		[DefaultValue(0)] public float EcosystemWeight;
 		[DefaultValue(0)] public float SpriteScale;
-		[DefaultValue(null)] public float[] Tint;
+		[DefaultValue(null)] public byte[] Tint;
 		[DefaultValue(true)] public bool Selectable;
 
 	}
